Piliz

+++ src/main/java/com/knowledge_hub/repository/CategoriesRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.knowledge_hub.entity.Categories;
 
-public interface CategoriesRepository extends JpaRepository<Categories, Integer>{
+public interface CategoriesRepository extends JpaRepository<Categories, Long>{
 
 }
